backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
aiofiles==23.2.1

# AI/ML
sentence-transformers==2.2.2
transformers==4.35.2
torch>=2.0.0
faiss-cpu>=1.9.0
pillow==10.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.0

# Utilities
requests==2.31.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

